+++ test/unit/helpers/fakes.ts
@@ -9,7 +9,8 @@ import type { SettingsSource } from '../../../src/host/settings'
 import { EN } from '../../../src/shared/l10n/en'
 import { BASE_LOCALE } from '../../../src/shared/l10n/text'
 import type { HostToWebviewMessage, SettingsSnapshot } from '../../../src/shared/protocol'
-import type { ChatSurface, WebviewHostContext } from '../../../src/host/views/webviewSetup'
+import type { ChatSurface } from '../../../src/host/views/chatSurface'
+import type { WebviewHostContext } from '../../../src/host/views/webviewSetup'
 import { EventEmitter, FakeUri, Uri } from '../mocks/vscode'
 
 function acceptMessage(_message: unknown): Thenable<boolean> {